tckgen: Fix linking issues using .cpp files and explicit inling
